requirements.txt
django==1.11.5
django-modeltranslation=0.12.2